backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.19
email-validator==2.2.0
slowapi==0.1.9
python-dotenv==1.0.1
requests==2.32.3
dnspython==2.7.0
python-whois==0.9.5
reportlab==4.2.5
cryptography==44.0.0
